+++ crates/ryo3-jiff/src/internal/datetime_round.rs
@@ -1,5 +1,4 @@
-use crate::dev::RyDateTimeRound;
-use crate::JiffUnit;
+use crate::{JiffUnit, RyDateTimeRound};
 use jiff::civil::DateTimeRound;
 use jiff::{SpanRound, ZonedRound};
 use pyo3::FromPyObject;
