+++ crates/ryo3-jiff/src/ry_datetime.rs
@@ -33,7 +33,7 @@ impl From<DateTime> for RyDateTime {
 impl RyDateTime {
     #[new]
     #[pyo3(signature = ( year, month, day, hour=0, minute=0, second=0, subsec_nanosecond=0))]
-    pub fn new(
+    pub fn py_new(
         year: i16,
         month: i8,
         day: i8,
