+++ crates/ryo3-std/src/lib.rs
@@ -14,9 +14,15 @@
 #![allow(clippy::unused_self)]
 
 use pyo3::prelude::*;
-mod duration;
-pub use duration::PyDuration;
+use time::duration;
+
+pub mod fs;
+pub mod time;
+
+pub use fs::*;
+pub use time::duration::PyDuration;
 pub fn pymod_add(m: &Bound<'_, PyModule>) -> PyResult<()> {
     duration::pymod_add(m)?;
+    fs::pymod_add(m)?;
     Ok(())
 }
