+++ crates/ryo3-std/src/time/mod.rs
@@ -1,0 +1,1 @@
+pub mod duration;
