+++ crates/ryo3/src/libs.rs
@@ -1,5 +1,38 @@
 use pyo3::prelude::*;
 
+// #[cfg(feature = "jiff")]
+// pub fn pymod_add_jiff(m: &Bound<'_, PyModule>) -> PyResult<()> {
+//     // classes
+//     m.add_class::<ryo3_jiff::RyDate>()?;
+//     m.add_class::<ryo3_jiff::RyDateTime>()?;
+//     m.add_class::<ryo3_jiff::RySignedDuration>()?;
+//     m.add_class::<ryo3_jiff::RySpan>()?;
+//     m.add_class::<ryo3_jiff::RyTime>()?;
+//     m.add_class::<ryo3_jiff::RyTimeZone>()?;
+//     m.add_class::<ryo3_jiff::RyTimestamp>()?;
+//     m.add_class::<ryo3_jiff::RyZoned>()?;
+//     m.add_class::<ryo3_jiff::RyOffset>()?;
+//
+//     // difference
+//     m.add_class::<ryo3_jiff::RyDateDifference>()?;
+//     m.add_class::<ryo3_jiff::RyDateTimeDifference>()?;
+//     m.add_class::<ryo3_jiff::RyTimeDifference>()?;
+//     m.add_class::<ryo3_jiff::RyTimestampDifference>()?;
+//
+//     // round
+//     m.add_class::<ryo3_jiff::RyDateTimeRound>()?;
+//
+//     // functions
+//     m.add_function(wrap_pyfunction!(ryo3_jiff::date, m)?)?;
+//     m.add_function(wrap_pyfunction!(ryo3_jiff::time, m)?)?;
+//     m.add_function(wrap_pyfunction!(ryo3_jiff::datetime, m)?)?;
+//     m.add_function(wrap_pyfunction!(ryo3_jiff::offset, m)?)?;
+//     m.add_function(wrap_pyfunction!(ryo3_jiff::timespan, m)?)?;
+//
+//
+//     ryo3_jiff::pymod_add(m)?;
+//     Ok(())
+// }
 pub fn pymod_add(m: &Bound<'_, PyModule>) -> PyResult<()> {
     #[cfg(feature = "brotli")]
     ryo3_brotli::pymod_add(m)?;
